benches for tree walks
